+++ example/lib/routes.dart
@@ -4,7 +4,7 @@ import 'package:awesome_notifications_example/pages/firebase_test_page.dart';
 
 import 'package:awesome_notifications_example/pages/media_details_page.dart';
 import 'package:awesome_notifications_example/pages/notification_details_page.dart';
-import 'package:awesome_notifications_example/pages/notification_examples_page.dart';
+import 'package:awesome_notifications_example/pages/home_page.dart';
 
 const String PAGE_HOME = '/';
 const String PAGE_MEDIA_DETAILS = '/media-details';
@@ -12,7 +12,7 @@ const String PAGE_NOTIFICATION_DETAILS = '/notification-details';
 const String PAGE_FIREBASE_TESTS = '/firebase-tests';
 
 Map<String, WidgetBuilder> materialRoutes = {
-  PAGE_HOME: (context) => NotificationExamplesPage(),
+  PAGE_HOME: (context) => HomePage(),
   PAGE_MEDIA_DETAILS: (context) => MediaDetailsPage(),
   PAGE_NOTIFICATION_DETAILS: (context) => NotificationDetailsPage(
         ModalRoute.of(context)!.settings.arguments as ReceivedNotification,